Split mask_rcnn.py into several files
