+++ torchvision/models/detection/__init__.py
@@ -1,1 +1,3 @@
+from .faster_rcnn import *
 from .mask_rcnn import *
+from .keypoint_rcnn import *
